Updated index.html with lots of sweet goodness
Removed the title attributes from links because they are useless. Useless for SEO or any other purpose. And they aren't any good for accessibility either: https://www.a11yproject.com/posts/title-attributes/

Modified the dfn attribute, which should really be wrapped around the word. The way the sentence is structured, parsers and decent algos will be able to comprehend word and its definition as it's written.

Replaced all of the' with' because gross.

Removed header element. If you're not using the head or body element, then why the fuck use the header element. Totally useless.

Removed aside element, because it's really part of the article. It's not a true aside and is therefore not being used semantically as intended. If you wanted to use an aside element, it would be used to reference something else that's vaguely related to the article. But this is completely relevant to the article.

Almost every paragraph didn't have a closing p element. What kind of sick bastard leaves their paragraph wide open for anyone to tickle?! Fixed!

Added a bunch of abbr elements to the acronyms and abbreviations. Unlike title attributes on links, these are actually useful. They're helpful to readers who aren't familiar with an acronym or abbreviation and they disambiguate the meaning for search engines (that's a good thing). 

Added canonical link. Kind of helpful considering this is a public repos and anyone could post it anywhere. You want that hint to be there for search bots to consider this domain the end all, be all domain.

Added a better meta for UTF-8 content type

All this talk on the page about HTTPS and SEO but no meta description. For shame!

Added Open Graph data because you want to be social, don't you? DON'T YOU!? I made an Open Graph image for it. Feel free to change it to whatever you want, but it needs to exist and should be no smaller than 1200x630 px. It's on IPFS, just for the heck of it. Let's live in the future, but feel free to add it to the repose if you want.

Added Schema structured data in JSON format. That's the format Google prefers and you should disambiguate ALL THE THINGS and talk them bots in their own language.

Added periods to the headings that didn't have them. Consistency would make mother proud.

Added system fonts so the main system font for the OS would be used. If that's not elegant as fuck, I don't know what is. It's instantaneous and nothing is loaded because it already IS loaded. It's also consistent with the paragraph about using fonts on the page.